Deleting a video from the index waits for response
So when we retry analysis we only start the upload & analysis after the
old asset was removed.

[#173509428]

+++ src/main/kotlin/com/boclips/videoanalyser/application/RetryVideoAnalysis.kt
@@ -20,7 +20,9 @@ class RetryVideoAnalysis(
         AnalyseVideo.logger.info { "Video $videoId received to retry analysis (language: ${request.language?.toLanguageTag() ?: "detect"})" }
 
         try {
-            videoAnalyserService.deleteVideo(videoId)
+            if (videoAnalyserService.isAnalysed(videoId)) {
+                videoAnalyserService.deleteVideo(videoId)
+            }
             analyseVideo.execute(VideoAnalysisRequested(videoId, request.videoUrl, request.language))
         } catch (e: Exception) {
             AnalyseVideo.logger.warn(e) { "Retry analysis of video $videoId failed and will not be retried." }
